Broke Groupable into two parts
