+++ lib/season_groupable.rb
@@ -1,5 +1,5 @@
-module Groupable
-  def games_by_team_id_and_season(season) #needs test
+module SeasonGroupable
+  def games_by_team_id_and_season(season)
     games_by_season = season_grouper[season] #season grouper is all games from the games csv grouped by season in arrays
     home_games = games_by_season.group_by { |game| game.home_team_id }
     away_games = games_by_season.group_by { |game| game.away_team_id }
@@ -12,16 +12,7 @@ module Groupable
     @games.group_by { |game| game.season }
   end
 
-  def all_team_games(team_id) #games helper, returns all of a team's games in an array
-    @games.find_all { |game| game.home_team_id == team_id || game.away_team_id == team_id }
-  end
-
-  def team_season_grouper(team_id) #helper, groups all of a team's games by season in a hash: the key is the season and the values are the team's games for that season
-    all_games = all_team_games(team_id)
-    all_games.group_by { |game| game.season }
-  end
-
- def games_by_season(season_id) #helper method
+ def games_by_season(season_id)
     game_id_list = []
     @games.each do |game|
       if game.season == season_id
@@ -31,9 +22,9 @@ module Groupable
     game_id_list
   end
 
-  def all_tackles_this_season(season) #needs test
+  def all_tackles_this_season(season)
     tackles_by_team = Hash.new(0)
-    @game_stats.games_by_team_id_and_season(season).flat_map { |team_id, games| games.map { |game| tackles_by_team[team_id] += 
+    @game_stats.games_by_team_id_and_season(season).flat_map { |team_id, games| games.map { |game| tackles_by_team[team_id] +=
       @game_teams_stats.number_of_tackles(team_id, game.game_id) }}
     tackles_by_team
   end
